pyproject.toml
[project]
name = "spotify-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "airflow-provider-duckdb>=0.2.0",
    "apache-airflow-providers-amazon>=9.13.0",
    "astronomer-cosmos>=1.10.2",
    "black>=25.1.0",
    "boto3>=1.40.31",
    "duckdb>=1.3.2",
    "flake8>=7.3.0",
    "load-dotenv>=0.1.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
    "ruff>=0.13.0",
    "uv>=0.8.17",
]
